+++ Utils/LoggerUtil.cs
@@ -3,17 +3,17 @@ using BepInEx.Logging;
 
 namespace MonitorLabels.Utils;
 
-public static class LoggerUtil //TODO: Add log levels to the config so people can choose what messages they see (flag?)
+public static class LoggerUtil
 {
 	private static ManualLogSource logger;
 
-	private static ConfigEntry<bool> configEnableLogger;
+	private static ConfigEntry<LogLevel> configLoggingLevel;
 
-	public static bool IsLoggerEnabled => configEnableLogger.Value;
+	public static bool IsLoggerEnabled => configLoggingLevel.Value > 0;
 
-	internal static void Initialize(ConfigEntry<bool> enableLoggerEntry, ManualLogSource logSource)
+	internal static void Initialize(ConfigEntry<LogLevel> enableLoggerEntry, ManualLogSource logSource)
 	{
-		configEnableLogger = enableLoggerEntry;
+		configLoggingLevel = enableLoggerEntry;
 
 		logger = logSource;
 	}
@@ -30,31 +30,61 @@ public static class LoggerUtil //TODO: Add log levels to the config so people ca
 	
 	internal static void LogMessage(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Message))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Message, data);
 	}
 
 	internal static void LogInfo(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Info))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Info, data);
 	}
 	
 	internal static void LogDebug(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Debug))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Debug, data);
 	}
 
 	internal static void LogError(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Error))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Error, data);
 	}
 
 	internal static void LogWarning(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Warning))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Warning, data);
 	}
 
 	internal static void LogFatal(object data)
 	{
+		if (!configLoggingLevel.Value.HasFlag(LogLevel.Fatal))
+		{
+			return;
+		}
+		
 		Log(LogLevel.Fatal, data);
 	}
 }