+++ ModPackage/CHANGELOG.md
@@ -4,6 +4,12 @@ Add an option to show a label on scrap (name or value or both, your choice)!
 
 Fix the labels for the MaskedPlayerEnemy (mimics)!
 
+Fix the target colour not working correctly if another player changes the radar focus
+
+Fix the label for a spider appearing rotated when the spider is crawling on the wall/ceiling
+
+Change 'EnableLogger' to a 'LogLevel' in the config, so you can now modify which logs will be logged
+
 # Version 1.0.0
 
 Add labels for Players, Radar boosters and enemies